refactor: 후처리기를 통한 proxy 생성 적용
