+++ src/main/java/springbook/proxy/TransactionAdvice.java
@@ -1,4 +1,4 @@
-package springbook.service;
+package springbook.proxy;
 
 import org.aopalliance.intercept.MethodInterceptor;
 import org.aopalliance.intercept.MethodInvocation;
