+++ src/main/java/springbook/proxy/hello/Hello.java
@@ -1,4 +1,4 @@
-package springbook.proxy;
+package springbook.proxy.hello;
 
 public interface Hello {
