+++ src/main/java/springbook/proxy/hello/HelloTarget.java
@@ -1,4 +1,4 @@
-package springbook.proxy;
+package springbook.proxy.hello;
 
 public class HelloTarget implements Hello {
