+++ src/main/java/springbook/proxy/hello/Uppercasehandler.java
@@ -1,4 +1,4 @@
-package springbook.proxy;
+package springbook.proxy.hello;
 
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
